requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Supabase
supabase==2.8.0
postgrest==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Gemini AI
google-generativeai==0.8.3

# Machine Learning & Recommendations
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
scipy==1.14.1
surprise==1.1.3
implicit==0.7.1

# Utilities
httpx==0.27.2
aiofiles==24.1.0
python-dateutil==2.9.0

# CORS
python-cors==1.0.0

